+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### v1.10.0
 
 #### Major
- - JsonPath function added. You can now use `.pick(<string>)` on `String`, `Hash` and `OpenStruct` to use JsonPath selection.
+ - JsonPath function added. You can now use `.pick(<string>)` on `String`, `Hash` and `OpenStruct` to use JsonPath selection. For more information on JsonPath see https://goessner.net/articles/JsonPath/
 
 #### Minor
  - HTTP logging of non JSON responses is fixed