+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Major
  - Helpers methods and functions added
    - `string.as_timestamp` is now available. Parses a string a unix timestamp
+   - `string.file_size` is now available. Interprets string as file path and gives back the file size if the file exists.
    - `now` returns the current time (like `Time.now`)
    - `array.last` method added. Returns the last element of list.
    - `Hash.default_to!` and `OpenStruct.default_to!` method added. Sets default values, when value in Hash is `nil`.